lecture/requirements.txt
# AI 워크플로우 강의 자료 생성 시스템 의존성
anthropic>=0.49.0
python-pptx>=1.0.0
Pillow>=10.0.0
feedparser>=6.0.0
python-docx>=1.1.0
